@@ -6,7 +6,7 @@
     "build": "vite build && liferay-npm-bundler-improved"
   },
   "dependencies": {
-    "react": "17.0.2"
+    "react": "18.2.0"
   },
   "devDependencies": {
     "@rollup/plugin-replace": "^3.0.0",
@@ -14,7 +14,7 @@
     "@types/react-dom": "^17.0.7",
     "@vitejs/plugin-react": "^2.0.0",
     "liferay-npm-bundler-improved": "workspace:^",
-    "react-dom": "17.0.2",
+    "react-dom": "18.2.0",
     "rollup": "^2.78.1",
     "tslib": "^2.3.1",
     "vite": "^3.0.4"
